+++ src/main/resources/com/github/jlangch/venice/setup/repl.sh
@@ -6,7 +6,7 @@
 #    |
 #    +-- libs
 #    |    +-- repl.json
-#    |    +-- venice-1.12.39.jar
+#    |    +-- venice-1.12.40.jar
 #    |
 #    +-- scripts
 #    |    +-- script-1.venice